Implemented gradients, magnitude and direction calculation on gpu Solved an issue with gradients calculation on gpu Implemented magnitude and direction calculation on gpu Minor changes

+++ utils/hog_utils.c
@@ -4,14 +4,14 @@
 //calcola la magnitude del gradiente
 void compute_magnitude(unsigned char *gradientX, unsigned char *gradientY, unsigned char *magnitude, int size){
   for (int i = 0; i < size; i++) {
-    magnitude[i] = sqrt(pow(gradientX[i], 2) + pow(gradientY[i], 2));
+    magnitude[i] = sqrtf(powf(gradientX[i], 2) + powf(gradientY[i], 2));
   }
 }
 
 //calcola la direzione del gradiente
 void compute_direction(unsigned char *gradientX, unsigned char * gradientY, unsigned char *direction, int size) {
 	for (int i = 0; i < size; i++) {
-    direction[i] = atan2(gradientY[i], gradientX[i]) * (180 / PI);
+    direction[i] = atan2f(gradientY[i], gradientX[i]) * (180 / PI);
   }
 }
